Move disruptive node tests into separate test fixtures
